Refined OOP of Buildings, added functionality
Buildings now use a Generate() function to do most of their construction, because this can be modified by descendent classes to allow for a great deal of customisation.
Buildings now take in a colour as a parameter, not sure if I think this is good or should be done internally to the constructor but works thus far.
Pros: We can set these colours based on the HUD
Cons: Not internal to the class
Buildings also have delete functions.