Added 2 new building variations, and sunlight
Highrise buildings have been put in their own file,
Farmhouses are another type of building sitting in the corner of a field.
These should act as good examples for further building type construction :)
Sunlight is a large directional light that cast shadows. If you have trouble with it you can disable it and just use the camera light for a while.
To angle the sunlight use the 'sunTargetAngle' vector.